Refactor dependency injection: move app_dependency_injection.dart to global_di and update imports in home_screen.dart and main.dart
