+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:ugaoo/core/environment/app_dependency_injection.dart';
+import 'package:ugaoo/core/global_di/app_dependency_injection.dart';
 import 'package:ugaoo/home_screen.dart';
 
 // Provided via: --dart-define=APP_FLAVOR=<dev|stg|prod>
